+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Blueprint for a citation generator 
 - MLA Citation rules to citation generator
 - ElasticSearch 7.10+ language per-field indexing & searching
+- Query parameter to manifest muse urls
 ### Fixed
 - Classify Performance improvements
 - Improved stability during clustering processes